mcp_servers/python/servers/QUICKBOOKS/pyproject.toml
[project]
name = "mcp-quickbooks"
version = "0.1.0"
description = "QuickBooks MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-dotenv",
    "python-quickbooks",
    "fastapi",
    "uvicorn",
    "requests"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-quickbooks = "mcp_quickbooks.__main__:main" 